pyproject.toml
[tool.poetry]
name = "tic-tac-toe-bot-project"
version = "0.1.0"
description = ""
authors = ["stas <chousetheone@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
python-telegram-bot = "^21.4"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
